# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: LittleAgResume API
  version: 0.2.0
  description: Contract for authentication and resume management.
servers:
  - url: http://localhost:8080
paths:
...
  /api/resumes:
    get:
      operationId: listResumes
      tags:
        - Resume
      summary: List the current user's resumes
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          schema:
            type: string
            maxLength: 80
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ResumeStatus"
        - name: sort
          in: query
          schema:
            $ref: "#/components/schemas/ResumeSort"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            enum: [6, 12, 24]
            default: 6
      responses:
        "200":
          description: Paginated resume summaries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createResume
      tags:
        - Resume
      summary: Create a blank draft resume
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResumeRequest"
      responses:
        "200":
          description: Created resume.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/resumes/stats:
    get:
      operationId: getResumeStats
      tags:
        - Resume
      summary: Get resume statistics for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user's aggregate resume statistics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeStatsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/resumes/import:
    post:
      operationId: importResume
      tags:
        - Resume
      summary: Import a LittleAgResume envelope as a new draft
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportResumeRequest"
      responses:
        "200":
          description: Imported resume.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/resumes/{resumeId}:
    parameters:
      - $ref: "#/components/parameters/ResumeId"
    get:
      operationId: getResume
      tags:
        - Resume
      summary: Get a resume owned by the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Resume details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      operationId: updateResume
      tags:
        - Resume
      summary: Update resume metadata or opaque content
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResumeRequest"
      responses:
        "200":
          description: Updated resume.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      operationId: deleteResume
      tags:
        - Resume
      summary: Delete a resume owned by the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Resume deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/resumes/{resumeId}/copy:
    parameters:
      - $ref: "#/components/parameters/ResumeId"
    post:
      operationId: copyResume
      tags:
        - Resume
      summary: Copy a resume as a new draft
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Copied resume.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/resumes/{resumeId}/import:
    parameters:
      - $ref: "#/components/parameters/ResumeId"
    put:
      operationId: replaceResumeImport
      tags:
        - Resume
      summary: Replace an existing resume from a LittleAgResume envelope
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ImportResumeRequest"
                - type: object
                  required:
                    - expectedRevision
                  properties:
                    expectedRevision:
                      type: integer
                      format: int64
                      minimum: 1
      responses:
        "200":
          description: Replaced resume.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/resumes/{resumeId}/export/pdf:
    parameters:
      - $ref: "#/components/parameters/ResumeId"
    post:
      operationId: exportResumePdf
      tags:
        - Resume
      summary: Render the resume to a PDF via headless Chrome; records the export on success
      security:
        - bearerAuth: []
      responses:
        "200":
          description: PDF bytes.
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/resumes/{resumeId}/print:
    parameters:
      - $ref: "#/components/parameters/ResumeId"
    get:
      operationId: getResumePrintData
      tags:
        - Resume
      summary: Print-page data endpoint authorized by a short-lived print token
      parameters:
        - name: X-Print-Token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resume detail plus inline avatar for printing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumePrintResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/resumes/{resumeId}/avatar:
    parameters:
      - $ref: "#/components/parameters/ResumeId"
    get:
      operationId: getResumeAvatar
      tags:
        - Resume
      summary: Get the current resume avatar
      security:
        - bearerAuth: []
      responses:
        "200":
          description: JPEG avatar bytes.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putResumeAvatar
      tags:
        - Resume
      summary: Replace the current resume avatar
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: A 500×700 JPEG portrait cropped to the standard 5:7 one-inch-photo ratio.
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Updated resume.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      operationId: deleteResumeAvatar
      tags:
        - Resume
      summary: Delete the current resume avatar
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated resume.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/resumes/{resumeId}/school-logo:
    parameters:
      - $ref: "#/components/parameters/ResumeId"
    get:
      operationId: getResumeSchoolLogo
      tags:
        - Resume
      summary: Get the current resume school logo
      security:
        - bearerAuth: []
      responses:
        "200":
          description: PNG school logo bytes.
          content:
            image/png:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putResumeSchoolLogo
      tags:
        - Resume
      summary: Replace the current resume school logo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: A 500×500 PNG school logo cropped to a square while preserving transparency.
        content:
          image/png:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Updated resume.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      operationId: deleteResumeSchoolLogo
      tags:
        - Resume
      summary: Delete the current resume school logo
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated resume.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResumeId:
      name: resumeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request, including ErrUsernameFormatInvalid and validation errors.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            ErrUsernameFormatInvalid:
              value:
                code: 101008
                message: Username format is invalid.
                data: null
    Unauthorized:
      description: Authentication failed, token missing, token expired, or token invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            ErrTokenExpired:
              value:
                code: 101003
                message: Token expired.
                data: null
            ErrTokenInvalid:
              value:
                code: 101004
                message: Token invalid.
                data: null
    Forbidden:
      description: The authenticated identity cannot perform this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Duplicate email or username, including ErrUsernameExists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            ErrUsernameExists:
              value:
                code: 101007
                message: Username already exists.
                data: null
    TooManyRequests:
      description: ErrTooManyRequests response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            ErrTooManyRequests:
              value:
                code: 100006
                message: Too many requests.
                data: null
    InternalServerError:
      description: Server error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: A required backend resource is temporarily unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    EmptyResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            code:
              type: integer
              format: int32
              enum:
                - 0
            message:
              type: string
              enum:
                - ""
            data:
              nullable: true
              example: null
...
    ResumeStatus:
      type: string
      enum:
        - draft
        - completed
    ResumeSort:
      type: string
      enum:
        - updated_desc
        - updated_asc
        - created_desc
        - title_asc
...
    ResumeDetail:
      allOf:
        - $ref: "#/components/schemas/ResumeSummary"
        - type: object
          required:
            - contentVersion
            - content
          properties:
            contentVersion:
              type: integer
              enum: [2, 3, 4]
            content:
              $ref: "#/components/schemas/ResumeContent"
...
    CreateResumeRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 80
    UpdateResumeRequest:
      type: object
      minProperties: 1
      required:
        - expectedRevision
      properties:
        expectedRevision:
          type: integer
          format: int64
          minimum: 1
        title:
          type: string
          minLength: 1
          maxLength: 80
        status:
          $ref: "#/components/schemas/ResumeStatus"
        profileAlignment:
          $ref: "#/components/schemas/ProfileAlignment"
        templateId:
          type: string
          nullable: true
          deprecated: true
          enum:
            - modern-editorial
            - classic-professional
        contentVersion:
          type: integer
          enum: [2, 3, 4]
        content:
          $ref: "#/components/schemas/ResumeContent"
    ImportResumeRequest:
      type: object
      required:
        - version
        - title
        - content
      properties:
        version:
          type: integer
          enum: [2, 3, 4]
        title:
          type: string
          minLength: 1
          maxLength: 80
        templateId:
          type: string
          nullable: true
          deprecated: true
          enum:
            - modern-editorial
            - classic-professional
        profileAlignment:
          $ref: "#/components/schemas/ProfileAlignment"
        avatar:
          type: string
          nullable: true
          maxLength: 750000
        schoolLogo:
          type: string
          nullable: true
          maxLength: 750000
        content:
          $ref: "#/components/schemas/ResumeContent"
...
    ResumeListResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          required:
            - data
          properties:
            code:
              type: integer
              format: int32
              enum: [0]
            message:
              type: string
              enum: [""]
            data:
              $ref: "#/components/schemas/ResumeListPayload"
    ResumeDetailResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          required:
            - data
          properties:
            code:
              type: integer
              format: int32
              enum: [0]
            message:
              type: string
              enum: [""]
            data:
              $ref: "#/components/schemas/ResumeDetail"
...
    ResumeStatsResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          required:
            - data
          properties:
            code:
              type: integer
              format: int32
              enum: [0]
            message:
              type: string
              enum: [""]
            data:
              $ref: "#/components/schemas/ResumeStats"
...